修复mybatis generator重复生成mapper.xml内容的问题

+++ mall-admin/src/main/java/com/macro/mall/service/impl/SmsFlashPromotionProductRelationServiceImpl.java
@@ -53,7 +53,7 @@ public class SmsFlashPromotionProductRelationServiceImpl implements SmsFlashProm
     }
 
     @Override
-    public int getCount(Long flashPromotionId, Long flashPromotionSessionId) {
+    public long getCount(Long flashPromotionId, Long flashPromotionSessionId) {
         SmsFlashPromotionProductRelationExample example = new SmsFlashPromotionProductRelationExample();
         example.createCriteria()
                 .andFlashPromotionIdEqualTo(flashPromotionId)
